+++ extensions/positron-jupyter-debugger/src/runtimeNotebookDebugAdapterFactory.ts
@@ -3,13 +3,15 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as path from 'path';
 import * as positron from 'positron';
 import * as vscode from 'vscode';
 import { log } from './extension.js';
 import { DebugCellManager } from './notebook.js';
 import { DisposableStore } from './util.js';
 import { JupyterRuntimeDebugAdapter } from './runtimeDebugAdapter.js';
-import { createDebuggerOutputChannel } from './runtime.js';
+
+const DEBUGGER_OUTPUT_CHANNEL_DESCRIPTOR = vscode.l10n.t('Debugger');
 
 // TODO: How do we handle reusing a debug adapter/session across cells?
 export class RuntimeNotebookDebugAdapterFactory implements vscode.DebugAdapterDescriptorFactory, vscode.Disposable {
@@ -83,10 +85,21 @@ export class RuntimeNotebookDebugAdapterFactory implements vscode.DebugAdapterDe
 
 	private createDebugAdapterOutputChannel(runtimeSession: positron.LanguageRuntimeSession): vscode.LogOutputChannel {
 		let outputChannel = this._outputChannelByRuntimeSessionId.get(runtimeSession.metadata.sessionId);
+
 		if (!outputChannel) {
-			outputChannel = this._disposables.add(createDebuggerOutputChannel(runtimeSession));
+			const runtimeName = runtimeSession.runtimeMetadata.runtimeName;
+			const sessionMode = runtimeSession.metadata.sessionMode;
+			let sessionTitle: string;
+			if (runtimeSession.metadata.notebookUri) {
+				sessionTitle = path.basename(runtimeSession.metadata.notebookUri.fsPath);
+			} else {
+				sessionTitle = sessionMode.charAt(0).toUpperCase() + sessionMode.slice(1);
+			}
+			const name = `${runtimeName}: ${DEBUGGER_OUTPUT_CHANNEL_DESCRIPTOR} (${sessionTitle})`;
+			outputChannel = this._disposables.add(vscode.window.createOutputChannel(name, { log: true }));
 			this._outputChannelByRuntimeSessionId.set(runtimeSession.metadata.sessionId, outputChannel);
 		}
+
 		return outputChannel;
 	}
